% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparer_donnees_bilan.R
\name{ajouter_prioritaires}
\alias{ajouter_prioritaires}
\title{Ajouter l'information sur les ouvrages prioritaires}
\usage{
ajouter_prioritaires(donnees_bilan, ouvrages_prioritaires)
}
\arguments{
\item{donnees_bilan}{tableau de données préparées obtenu avec la fonction
\code{\link[=selectionner_donnees_bilan]{selectionner_donnees_bilan()}}}

\item{ouvrages_prioritaires}{tableau de données contenant au moins les deux
champs suivants: identifiant_roe (texte) et prioritaire (logique:
TRUE/FALSE).}
}
\value{
un tableau de données contenant les informations de l'export du ROE
complété par le statut prioritaire ou non des ouvrages.
}
\description{
Cette fonction permet d'ajouter une information sur le fait que les ouvrages
sont considérés ou non comme prioritaires au titre de la politique de
restauration de la continuité écologique.
}
\details{
Dans le tableau ouvrages_prioritaires, le champ identifiant_roe correspond
aux identifiants utilisés dans le tableau donnees_bilan. Le tableau peut ne
contenir que les ouvrages prioritaires, dans ce cas les ouvrages présents
dans donnees_bilan et absents du tableau ouvrages_prioritaires seront alors
automatiquement considérés comme non prioritaires.
}
