% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_completude.R
\name{cartographier_completude}
\alias{cartographier_completude}
\title{Cartographier la complétude des informations obligatoires et complémentaires}
\usage{
cartographier_completude(
  donnees_carte,
  reseau_hydro = NULL,
  listes2 = NULL,
  limites_zone = NULL,
  taille_obstacles = 2.5
)
}
\arguments{
\item{reseau_hydro}{données spatiales (classe \code{sf}) du réseau hydrographique.
Par défaut, n'est pas affiché (NULL)}

\item{listes2}{données spatiales (classe \code{sf}) de la partie du réseau.Par
défaut, n'est pas affiché (NULL) hydrographique classé en liste 2}

\item{limites_zone}{données spatiales (classe \code{sf}) des limites de la zone
représentée (région, département...). Par défaut, n'est pas affiché (NULL)}

\item{taille_obstacles}{taille des figurés représentant les obstacles
(défaut: 2.5)}

\item{donnees_carto}{données obtenues avec la fonction \code{\link[=preparer_donnees_carte_completude]{preparer_donnees_carte_completude()}}}
}
\description{
Localise sur une carte les obstacles non gelés et pour lesquels on dispose au
moins des coordonnées géographiques. La symbologie utilisée permet
d'identifier les ouvrages considérés comme prioritaires ainsi que le type
d'information manquante parmi les trois informations obligatoires (nom, type
et état de l'ouvrage) et deux informations complémentaires (classe de hauteur
de chute et équipement en dispositif de franchissement piscicole).
}
