% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_completude.R
\name{evaluer_completude}
\alias{evaluer_completude}
\title{Evaluer la complétude des données obligatoires et facultatives}
\usage{
evaluer_completude(donnees_bilan)
}
\arguments{
\item{donnees_bilan}{sélection des données exportées utilisées pour réaliser
des bilans, obtenue avec la fonction \code{\link[=preparer_donnees_bilan]{preparer_donnees_bilan()}}.}
}
\description{
Cette fonction permet d'évaluer pour chaque ouvrage la complétion de
l'information pour les champs obligatoires (nom, type, état et coordonnées)
et deux champs complémentaires (classe de hauteur de chute et équipement en
dispositif de franchissement piscicole) du ROE.
}
\details{
Il est vérifié qu'une information est fournie pour les champs nom_principal,
type_code, etat_code, pour les deux champs de coordonnées x_l93 et y_l93,
pour le champ hauteur_chute_etiage_classe ainsi que pour au moins un des
champs correspondant aux dispositifs de franchissement piscicole (fpi_code1 à
fpi_code5). Le nombre d'information obligatoire (les coordonnées sont
considérées comme une information unique) et complémentaire renseignées sont
également calculés.
}
