% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_cartes_obstacles.R
\name{generer_cartes_obstacles}
\alias{generer_cartes_obstacles}
\title{Générer des cartes de localisation des ouvrages ROE}
\usage{
generer_cartes_obstacles(
  donnees,
  codes_roe = NULL,
  dossier_sortie = getwd(),
  sous_dossiers = NULL,
  fond_carte
)
}
\arguments{
\item{donnees}{données issus d'un export ROE}

\item{codes_roe}{vecteur de codes ROE des obstacles pour lesquels on veut
générer des fiches terrain.}

\item{dossier_sortie}{chemin du dossier dans lequel seront enregistrées les
fiches terrain générées.}

\item{sous_dossiers}{vecteur contenant les noms de colonnes du tableau
\code{donnees} qui serviront à hiérarchiser les enregistrements des fiches (e.g.
par département en spécifiant "dept_nom")}

\item{fond_carte}{fond de carte, objet raster de classe stars}
}
\description{
Générer des cartes de localisation des ouvrages ROE
}
