% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_all_levels}
\alias{get_all_levels}
\title{Récupérer les valeurs uniques non nulles d'une colonne dans un tableau}
\usage{
get_all_levels(data, column)
}
\arguments{
\item{data}{tableau de données}

\item{column}{nom de la colonne pour laquelle on veut récupérer les valeurs
uniques non nulles}
}
\description{
Récupérer les valeurs uniques non nulles d'une colonne dans un tableau
}
