% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_donnees.R
\name{importer_donnees}
\alias{importer_donnees}
\title{Importer les données de l'export}
\usage{
importer_donnees(chemin_archive, bassins = NULL, departements = NULL)
}
\arguments{
\item{chemin_archive}{texte. Chemin vers l'archive zip contenant l'export ROE}

\item{bassins}{vecteur de texte. Noms des bassins hydrographiques (en
majuscules) à importer. Par défaut (NULL), tous les bassins sont importés.
Les obstacles avec un bassin non renseigné (NULL, NA ou '') ne sont pas
importés.}

\item{departements}{vecteur de texte. Codes INSEE des départements à
importer. Par défaut (NULL), tous les départements sont importés. Les
obstacles avec un département non renseigné (NULL, NA ou '') ne sont pas
importés.}
}
\value{
un tableau contenant l'ensemble de l'export ROE renseigné, filtré le
cas échéant par bassin hydrographique ou département.
}
\description{
Importer les données de l'export
}
