% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparer_donnees_bilan.R
\name{preparer_donnees_bilan}
\alias{preparer_donnees_bilan}
\title{Préparer les données pour faire le bilan}
\usage{
preparer_donnees_bilan(
  donnees_brutes,
  ouvrages_prioritaires = NULL,
  masses_eau = NULL
)
}
\arguments{
\item{donnees_brutes}{tableau de données obtenu avec la fonction
\code{\link[=importer_donnees]{importer_donnees()}}}

\item{ouvrages_prioritaires}{tableau de données contenant au moins les deux
champs suivants: identifiant_roe (texte) et prioritaire (logique:
TRUE/FALSE).}
}
\value{
un tableau de données contenant les informations de l'export du ROE
préparées et complétées par le statut prioritaire ou non des ouvrages.
}
\description{
Cette fonction est juste un 'wrapper' d'autres fonctions du package pour
réaliser en une étape la sélection des champs de l'export brut et l'ajout de
informations sur le caractère prioritaire ou non des ouvrages dans le cadre
des politiques de restauration de la continuité écologique.
}
