% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_completude.R
\name{preparer_donnees_carte_completude}
\alias{preparer_donnees_carte_completude}
\title{Préparer les données cartographiques résumant la complétude des informations}
\usage{
preparer_donnees_carte_completude(donnees_bilan)
}
\description{
Prépare un tableau spatialisé résumant pour chaque obstacle le nombre
d'informations obligatoires et complémentaires manquantes.
}
