% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_validation.R
\name{preparer_donnees_sankey}
\alias{preparer_donnees_sankey}
\title{Mettre un tableau de données dans un format adapté à la production de
diagramme de Sankey avec {ggalluvial}}
\usage{
preparer_donnees_sankey(donnees, etape, individu, groupe, ...)
}
\arguments{
\item{donnees}{tableau à mettre en forme}

\item{etape}{nom de la colonne correspondant à la variable selon laquelle on
veut suivre les évolutions (dates, ...)}

\item{individu}{nom de la colonne correspondant aux identifiants uniques des
'individus'}

\item{groupe}{nom de la colonne correspondant aux groupes auxquels les
individus appartiennent. L'appartenance d'un individu à un groupe peut
changer en fonction de \code{x}}

\item{...}{noms des colonnes pour lesquels on veut détailler les flux
(sous-ensembles des données)}
}
\description{
Le format de données requis pour construire un diagramme de Sankey est
particulier puisqu'il doit permettre de suivre l'appartenance d'individus à
des groupes entre plusieurs étapes. Cette fonction permet de gérer le
formatage d'un tableau de données au format 'tidy' pour pouvoir être utilisé
pour produire un diagramme de Sankey.
}
