% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{read_from_zip}
\alias{read_from_zip}
\title{Lire le contenu d'une archive}
\usage{
read_from_zip(zipfile, file, fun, ...)
}
\arguments{
\item{zipfile}{texte. Chemin de l'archive}

\item{file}{texte. Tout ou partie du nom du fichier à extraire}

\item{fun}{nom de la fonction à utiliser pour lire le fichier}

\item{...}{paramètres supplémentaires à passer à la fonction \code{fun}}
}
\description{
Lire le contenu d'une archive
}
