% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_completude.R
\name{synthetiser_completude}
\alias{synthetiser_completude}
\title{Synthétiser l'information sur la complétude des données obligatoires et
complémentaires}
\usage{
synthetiser_completude(donnees_bilan, ...)
}
\arguments{
\item{...}{noms des champs (sans "") permettant de définir les groupes pour
lesquels on veut le détail}
}
\value{

}
\description{
L'évaluation de la complétude des informations à l'ouvrage est synthétisée en
comptant le nombre d'ouvrages pour lequel il manque au moins une information
obligatoire (nom, type, état et coordonnées) et/ou au moins une information
facultative parmi les deux considérées (classe de hauteur de chute et
équipement en dispositif de franchissement piscicole).
}
\details{
Le décompte des ouvrages non gelés pour lesquels de l'information est
manquante peut être réalisé au niveau du jeu de données entier (par défaut),
ou au niveau de sous-ensembles en indiquant comme paramètres supplémentaires
(\code{...}), les noms de champs pour lesquels on veut regarder en détail (e.g.
\code{dept_nom} pour un détail par département, \code{prioritaire} pour un détail pour
les ouvrages prioritaires ou non). Fournir plusieurs noms de champs à la
place de \code{...} permet de considérer des groupes et sous-groupes définis par
les combinaisons de ces champs.
}
