% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_donnees.R
\name{telecharger_limites_region}
\alias{telecharger_limites_region}
\title{Télécharger les limites des régions administratives}
\usage{
telecharger_limites_region(codes_region = NULL)
}
\arguments{
\item{codes_region}{vecteur contenant les codes Insee des régions pour
lesquelles on veut les limites (par défaut, elles sont toutes retournées)}
}
\description{
Télécharger les limites des régions administratives
}
