% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_donnees.R
\name{telecharger_listes2}
\alias{telecharger_listes2}
\title{Télécharger les linéaires de cours d'eau classés en liste 2 au titre de la
continuité écologique (flux Sandre)}
\usage{
telecharger_listes2(limites_zone = NULL)
}
\arguments{
\item{limites_zone}{couche spatiale de type (multi)polygone délimitant l'aire
d'intérêt.}
}
\value{

}
\description{
Télécharger les linéaires de cours d'eau classés en liste 2 au titre de la
continuité écologique (flux Sandre)
}
