% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_donnees.R
\name{telecharger_masses_eau}
\alias{telecharger_masses_eau}
\title{Télécharger les limites des masses d'eau (France métropolitaine; flux Sandre)}
\usage{
telecharger_masses_eau(limites_zone = NULL)
}
\arguments{
\item{limites_zone}{couche spatiale de type (multi)polygone délimitant l'aire
d'intérêt.}
}
\description{
Les masses d'eau considérées correspondent aux
\href{https://id.eaufrance.fr/ddd/MDO/1.4/BVSpeMasseDEauSurface}{bassins versants spécifiques} tels que définis dans le Référentiel des masses d'eau. Si une
zone d'intérêt est fournie avec l'argument \code{limites_zone} (par exemple les
limites des régions administratives obtenues avec la fonction
\code{\link[=telecharger_limites_region]{telecharger_limites_region()}}), seule les masses d'eau de cette zone sont
retournées, sinon toutes les masses d'eau de France métropolitaine sont
retournées.
}
