% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_completude.R
\name{visualiser_completude}
\alias{visualiser_completude}
\title{Visualisation de la synthèse de la complétude des données obligatoires et
complémentaires}
\usage{
visualiser_completude(
  donnees_bilan,
  groupe = NULL,
  visualiser_prioritaires = FALSE,
  nombre_lignes = NULL,
  nombre_colonnes = NULL,
  ajuster_ymax = 1.1
)
}
\arguments{
\item{donnees_bilan}{sélection des données exportées utilisées pour réaliser
des bilans, obtenue avec la fonction \code{\link[=preparer_donnees_bilan]{preparer_donnees_bilan()}}.}

\item{groupe}{nom de la colonne pour laquelle on veut voir le détail}

\item{visualiser_prioritaires}{valeur logique (TRUE/FALSE) contrôlant si le
détail du nombre d'ouvrage est afficher en différenciant les ouvrages
prioritaires ou pas}

\item{nombre_lignes}{nombre de lignes sur lesquelles répartir les graphiques
crées pour chacun des \code{groupe}s. Calculé automatiquement par défaut}

\item{nombre_colonnes}{nombre de colonnes sur lesquelles répartir les
graphiques crées pour chacun des \code{groupe}s. Calculé automatiquement par
défaut.}

\item{ajuster_ymax}{valeur numérique par laquelle le maximum de l'axe y est
ajusté pour augmenter l'espace entre les barres et le titre. Valeur par
défaut de 1.1}
}
\description{
Représente sous la forme d'une graphique de type 'upset' la complétude des
informations obligatoires (nom, type, état et coordonnées) et de deux
informations complémentaires (classe de hauteur de chute et équipement en
dispositif de franchissement piscicole). Le nombre d'ouvrage non gelés
correspondant à chaque combinaison d'information manquante est représenté.
Cette répartition peut être déclinée par une variable de regroupement des
ouvrages (e.g. par département)
}
