% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiser_evolution_roe.R
\name{visualiser_evolution_roe}
\alias{visualiser_evolution_roe}
\title{Visualiser l'évolution du nombre d'ouvrages validés dans le ROE}
\usage{
visualiser_evolution_roe(
  donnees_bilan,
  groupe = NULL,
  log_y = FALSE,
  ajuster_xmax = 1.015
)
}
\arguments{
\item{donnees_bilan}{sélection des données exportées utilisées pour réaliser
des bilans, obtenue avec la fonction \code{\link[=preparer_donnees_bilan]{preparer_donnees_bilan()}}.}

\item{log_y}{valeur logique (TRUE/FALSE) pour exprimer ou non l'axe des y
(nombre d'obstacles validés) avec une échelle logarithmique}

\item{ajuster_xmax}{valeur numérique par laquelle le maximum de l'axe x est
ajusté pour laisser de la place aux annotations avec le nombre d'ouvrages.}
}
\value{
un objet ggplot
}
\description{
Cette fonction permet de visualiser simplement l'évolution du nombre cumulé
d'ouvrages dans le ROE. Cette évolution peut être déclinée par une variable
de regroupement des ouvrages (e.g. par département).
}
