% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_validation.R
\name{visualiser_evolution_validations}
\alias{visualiser_evolution_validations}
\title{Visualiser l'évolution du nombre d'obstacles en fonction de leur statut de
validation et de la complétude des informations}
\usage{
visualiser_evolution_validations(
  liste_bilans,
  ...,
  log_y = FALSE,
  text_size = 3
)
}
\arguments{
\item{liste_bilans}{une liste de tableaux obtenus avec la fonction
\code{\link[=preparer_donnees_bilan]{preparer_donnees_bilan()}}. Cette liste doit être nommée (e.g. avec les
dates des exports correspondants aux différents tableaux de bilan) et les
éléments de cette liste classés dans l'ordre dans lequel on veut les
afficher.}

\item{...}{noms des colonnes pour lesquels on veut détailler les flux
(sous-ensembles des données)}

\item{log_y}{valeur logique (TRUE/FALSE): les comptes d'individus doivent-ils
être exprimés en log ou non. Peut être utile pour suivre l'évolution des
groupes à faibles effectifs}

\item{text_size}{taille du texte affichant les nombres d'individu par groupe}
}
\description{
Cette fonction utilise des bilans réalisés à différentes dates pour suivre
l'évolution du statut de validation et de complétude de l'information des
obstacles.
}
