% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_validation.R
\name{visualiser_sankey}
\alias{visualiser_sankey}
\title{Visualiser les changements de groupes d'individus en fonction d'étapes à
l'aide d'un diagramme de Sankey}
\usage{
visualiser_sankey(df, etape, groupe, text_size = 3, log_y = FALSE, ...)
}
\arguments{
\item{df}{tableau mis en forme à l'aide de la fonction
\code{\link[=preparer_donnees_sankey]{preparer_donnees_sankey()}}}

\item{etape}{nom de la colonne correspondant à la variable selon laquelle on
veut suivre les évolutions (dates, ...)}

\item{groupe}{nom de la colonne correspondant aux groupes auxquels les
individus appartiennent. L'appartenance d'un individu à un groupe peut
changer en fonction de \code{x}}

\item{text_size}{taille du texte affichant les nombres d'individu par groupe}

\item{log_y}{valeur logique (TRUE/FALSE): les comptes d'individus doivent-ils
être exprimés en log ou non. Peut être utile pour suivre l'évolution des
groupes à faibles effectifs}

\item{...}{noms des colonnes pour lesquels on veut détailler les flux
(sous-ensembles des données)}
}
\description{
Visualiser les changements de groupes d'individus en fonction d'étapes à
l'aide d'un diagramme de Sankey
}
