% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluer_validation.R
\name{visualiser_validation}
\alias{visualiser_validation}
\title{Visualisation de la synthèse des validations}
\usage{
visualiser_validation(donnees_bilan, groupe = NULL)
}
\arguments{
\item{donnees_bilan}{sélection des données exportées utilisées pour réaliser
des bilans, obtenue avec la fonction \code{\link[=preparer_donnees_bilan]{preparer_donnees_bilan()}}.}

\item{groupe}{nom de la colonne pour laquelle on veut voir le détail}
}
\description{
Représente sous la forme d'une treemap, la répartition des nombres d'ouvrages
en fonction de leur statut de validation (Validé, Non validé ou Gelé). Cette
répartition peut être déclinée par une variable de regroupement des ouvrages
(e.g. par département)
}
