% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparer_donnees_bilan.R
\name{selectionner_donnees_bilan}
\alias{selectionner_donnees_bilan}
\title{Sélectionne les champs de l'export qui seront utilisés pour réaliser le bilan}
\usage{
selectionner_donnees_bilan(donnees_brutes, type = "roe")
}
\arguments{
\item{donnees_brutes}{tableau de données obtenu avec la fonction
\code{\link[=importer_donnees]{importer_donnees()}}}
}
\value{
un tableau de données ne contenant que la sélection de champs de
l'export qui sont utilisés par les fonctions de {BilanROE}
}
\description{
Sélectionne les champs de l'export qui seront utilisés pour réaliser le bilan
}
